updated some functions so they use error.h

+++ src/error/return_msg.c
@@ -6,13 +6,14 @@
 /*   By: orudek <orudek@student.42madrid.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/02 14:51:25 by orudek            #+#    #+#             */
-/*   Updated: 2023/10/02 14:53:42 by orudek           ###   ########.fr       */
+/*   Updated: 2023/10/02 15:09:39 by orudek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "error.h"
+#include "unistd.h"
+#include "libft.h"
 
-size_t return_msg(char *msg, int fd, size_t ret_val)
+unsigned long	return_msg(char *msg, int fd, unsigned long ret_val)
 {
 	if (msg)
 	{
