+++ src/error/return_perror.c
@@ -6,13 +6,13 @@
 /*   By: orudek <orudek@student.42madrid.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/02 14:51:31 by orudek            #+#    #+#             */
-/*   Updated: 2023/10/02 14:55:40 by orudek           ###   ########.fr       */
+/*   Updated: 2023/10/02 15:09:42 by orudek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "error.h"
+#include <stdio.h>
 
-size_t return_perror(char *msg, size_t ret_val)
+unsigned long	return_perror(char *msg, unsigned long ret_val)
 {
 	perror(msg);
 	return (ret_val);
